# core/aggregator/api/swagger.yaml
openapi: 3.0.0
info:
  title: Swagger Gandalf
  description: "This is a sample Petstore server.  You can find \nout more about Swagger\
    \ at \n[http://swagger.io](http://swagger.io) or on \n[irc.freenode.net, #swagger](http://swagger.io/irc/).\n"
  termsOfService: http://swagger.io/terms/
  contact:
    email: romain.fairant@orness.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0-oas3
...
servers:
- url: https://virtserver.swaggerhub.com/ditrit/Gandalf/1.0.0
- url: http://virtserver.swaggerhub.com/ditrit/Gandalf/1.0.0
...
paths:
...
  /user:
    get:
      tags:
      - user
      summary: List users
      operationId: listUser
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_8'
                x-content-type: application/json
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    post:
      tags:
      - user
      summary: Create a user
      operationId: createUser
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateUser'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /user/{userId}:
    get:
      tags:
      - user
      summary: Find user by ID
      description: Returns a single user
      operationId: getUserById
      parameters:
      - name: userId
        in: path
        description: ID of user to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    put:
      tags:
      - user
      summary: Updates a role
      operationId: updateUser
      parameters:
      - name: userId
        in: path
        description: ID of role that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        $ref: '#/components/requestBodies/createOrUpdateUser'
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Invalid ID supplied | Invalid request payload
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
    delete:
      tags:
      - user
      summary: Deletes a role
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        description: Role id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Succes
        "400":
          description: Invalid ID supplied
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /user/{userName}:
    get:
      tags:
      - user
      summary: Find user by Name
      description: Returns a single User
      operationId: getUserByName
      parameters:
      - name: userName
        in: path
        description: Name of user to return
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "404":
          description: User not found
        "500":
          description: Database error | Tenant not found
      security:
      - bearerAuth: []
  /user/login:
    post:
      tags:
      - user
      summary: Logs user into the system
      operationId: loginUser
      requestBody:
        $ref: '#/components/requestBodies/loginUser'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                x-content-type: application/json
        "400":
          description: Invalid request payload
        "401":
          description: Invalid username/password supplied
...
components:
  schemas:
    Model:
      type: object
      properties:
        id:
          type: integer
          description: The ID.
        createdAt:
          type: string
          description: The created date.
          format: date-time
        updatedAt:
          type: string
          description: The updated date.
          format: date-time
        deletedAt:
          type: string
          description: The deleted date.
          format: date-time
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        deletedAt: 2000-01-23T04:56:07.000+00:00
        id: 0
        updatedAt: 2000-01-23T04:56:07.000+00:00
    User:
      type: object
      properties:
        model:
          $ref: '#/components/schemas/Model'
        name:
          type: string
          description: The user name.
        email:
          type: string
          description: The user email.
        password:
          type: string
          description: The user password.
      example:
        password: password
        name: name
        model:
          createdAt: 2000-01-23T04:56:07.000+00:00
          deletedAt: 2000-01-23T04:56:07.000+00:00
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
        email: email
...
    inline_response_200_8:
      type: object
      properties:
        schema:
          $ref: '#/components/schemas/User'
      example:
        schema:
          password: password
          name: name
          model:
            createdAt: 2000-01-23T04:56:07.000+00:00
            deletedAt: 2000-01-23T04:56:07.000+00:00
            id: 0
            updatedAt: 2000-01-23T04:56:07.000+00:00
          email: email
...
  requestBodies:
...
    createOrUpdateUser:
      description: The user to create.
      content:
        application/json:
          schema:
            required:
            - email
            - name
            - password
            type: object
            properties:
              name:
                type: string
              email:
                type: string
              password:
                type: string
    loginUser:
      description: The user to create.
      content:
        application/json:
          schema:
            required:
            - email
            - password
            type: object
            properties:
              email:
                type: string
              password:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT